Refactored code to SOLID principles
